+++ src/test/java/com/epam/reportportal/utils/formatting/MarkdownUtilsTest.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2019 EPAM Systems
+ * Copyright 2024 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.epam.reportportal.utils.markdown;
+package com.epam.reportportal.utils.formatting;
 
+import com.epam.reportportal.utils.formatting.MarkdownUtils;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -22,7 +23,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import static com.epam.reportportal.utils.markdown.MarkdownUtils.*;
+import static com.epam.reportportal.utils.formatting.MarkdownUtils.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;